% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{draft_names_crosswalk}
\alias{draft_names_crosswalk}
\title{Eliteprospects draft list}
\format{
\subsection{\code{draft_names_crosswalk}}{

A data frame with 41 rows and 2 columns:
\describe{
\item{draft_name}{official name of the draft}
\item{link_component}{link embedded variable to demark particular draft}
}
}
}
\source{
\url{https://www.eliteprospects.com/draft/nhl-entry-draft}
}
\usage{
draft_names_crosswalk
}
\description{
A crosswalk of league drafts to the link suffixes for easy website creation
and lookup
}
\keyword{datasets}
