% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{eliteprospects_api}
\alias{eliteprospects_api}
\title{EliteProspects Caller}
\usage{
eliteprospects_api(path, params = list())
}
\arguments{
\item{path}{Path Components of API call}

\item{params}{Filter or other parameters of API call}
}
\value{
The response from the API call built with path and params
}
\description{
EliteProspects Caller
}
