% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get API Key}
\usage{
get_api_key()
}
\value{
API key set by set_api_key
}
\description{
Returns the API key entered in set_api_key.
}
\seealso{
\code{\link{set_api_key}} for setting the API key

Other API functions: \code{\link{set_api_key}}
}
