% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliteprospects_api.R
\name{get_user_agent}
\alias{get_user_agent}
\title{Get the User Agent String}
\usage{
get_user_agent()
}
\value{
the user agent string. Set the httr::user_agent with this string
}
\description{
Get the set User Agent string for building calls.
}
