% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{getCountry}
\alias{getCountry}
\title{Get a specific Country Entry}
\usage{
getCountry(id = NULL, filters = NULL, sorts = NULL, fields = NULL,
  limit = 0, offset = 0)
}
\arguments{
\item{id}{The transfer entry ID to get}

\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}
}
\value{
The retired number entry from the API
}
\description{
Get a specific Country Entry
}
