% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayerData}
\alias{getPlayerData}
\title{Get A Player's Data}
\usage{
getPlayerData(player_id = NULL)
}
\arguments{
\item{player_id}{The player_id from Elite Prospects of the player you wish to retireve}
}
\value{
The player's information from Elite Prospects, filtered if appropriate
}
\description{
Get A Player's Data
}
