% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_frontend.R
\name{getPlayerProfile}
\alias{getPlayerProfile}
\title{Get Player Profile}
\usage{
getPlayerProfile(player_id = NULL, filters = NULL, sorts = NULL,
  fields = NULL, limit = 0, offset = 0)
}
\arguments{
\item{player_id}{Player whos stats you wish to pull}

\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}
}
\value{
The players stats from the API
}
\description{
Get Player Profile
}
