% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{league_names_crosswalk}
\alias{league_names_crosswalk}
\title{Eliteprospects league list}
\format{
\subsection{\code{league_names_crosswalk}}{

A data frame with 482 rows and 3 columns:
\describe{
\item{full_name}{full name of the league (National Hockey League, etc.)}
\item{abbreviation}{abbreviation of the league (NHL, etc.)}
\item{link_component}{link embedded variable to demark particular league}
}
}
}
\source{
\url{https://www.eliteprospects.com/leagues}
}
\usage{
league_names_crosswalk
}
\description{
Eliteprospects league list
}
\keyword{datasets}
