% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues.R
\name{listLeagues}
\alias{listLeagues}
\title{List Leagues}
\usage{
listLeagues(filters = NULL, sorts = NULL, fields = NULL, limit = 0,
  offset = 0)
}
\arguments{
\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}
}
\value{
The leagues from the API
}
\description{
List Leagues
}
