% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_frontend.R
\name{listRetiredNumbers}
\alias{listRetiredNumbers}
\title{List Retired Numbers}
\usage{
listRetiredNumbers(filters = NULL, sorts = NULL, fields = NULL,
  limit = 0, offset = 0)
}
\arguments{
\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}
}
\value{
The retired numbers from the API
}
\description{
List Retired Numbers
}
