% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_fields}
\alias{process_fields}
\title{Process fields}
\usage{
process_fields(fields)
}
\arguments{
\item{sorts}{The field(s) to process}
}
\value{
properly formatted field(s)
}
\description{
Process fields
}
