% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_filters}
\alias{process_filters}
\title{Process Filters}
\usage{
process_filters(filters)
}
\arguments{
\item{filters}{The filter(s) to process}
}
\value{
properly formatted filter(s)
}
\description{
Process Filters
}
