% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_limit}
\alias{process_limit}
\title{Process limit}
\usage{
process_limit(limit)
}
\arguments{
\item{limit}{The limit to process}
}
\value{
properly formatted limit
}
\description{
Process limit
}
