% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_minScore}
\alias{process_minScore}
\title{Process minScore}
\usage{
process_minScore(minScore)
}
\arguments{
\item{minScore}{The minScore to process}
}
\value{
properly formatted offset
}
\description{
Process minScore
}
