% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_offset}
\alias{process_offset}
\title{Process offset}
\usage{
process_offset(offset)
}
\arguments{
\item{offset}{The offset to process}
}
\value{
properly formatted offset
}
\description{
Process offset
}
