% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_params}
\alias{process_params}
\title{Process Parameters for API call building}
\usage{
process_params(filters = NULL, sorts = NULL, fields = NULL, limit = 0,
  offset = 0)
}
\arguments{
\item{filters}{filter(s) to process, or NULL}

\item{sorts}{sort(s) to process, or NULL}

\item{limit}{limit to process, or None}

\item{offset}{offset to process, or None}
}
\value{
properly formatted parameters
}
\description{
Process Parameters for API call building
}
