% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_search_params}
\alias{process_search_params}
\title{Process Parameters for API searches}
\usage{
process_search_params(types = NULL, filters = NULL, sorts = NULL,
  fields = NULL, limit = 0, offset = 0, minScore = 0)
}
\arguments{
\item{types}{type(s) to process, or NULL}

\item{filters}{filter(s) to process, or NULL}

\item{sorts}{sort(s) to process, or NULL}

\item{fields}{field(s) to process, or NULL}

\item{limit}{limit to process, or None}

\item{offset}{offset to process, or None}

\item{minScore}{minScore to process, or None}
}
\value{
properly formatted parameters
}
\description{
Process Parameters for API searches
}
