% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_sorts}
\alias{process_sorts}
\title{Process sorts}
\usage{
process_sorts(sorts)
}
\arguments{
\item{sorts}{The sort(s) to process}
}
\value{
properly formatted sort(s)
}
\description{
Process sorts
}
