% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{process_types}
\alias{process_types}
\title{Process types}
\usage{
process_types(types)
}
\arguments{
\item{types}{The type(s) to process}
}
\value{
properly formatted type(s)
}
\description{
Process types
}
