% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_EP}
\alias{search_EP}
\title{Search for an item in EliteProspects}
\usage{
search_EP(query = NULL, types = NULL, filters = NULL, sorts = NULL,
  fields = NULL, limit = 0, offset = 0, minScore = 0)
}
\arguments{
\item{query}{The query you wish to search for}

\item{types}{Any type of record to limit search to (eg. player, staff) as vector}

\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve as vector}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}

\item{minScore}{Minimum search score for returned results.}
}
\value{
The search results
}
\description{
Search for an item in EliteProspects
}
