% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_persistent.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key}
\usage{
set_api_key(new_api_key = "DEFAULT_API")
}
\arguments{
\item{new_api_key}{API key string from eliteprospects.}
}
\description{
API keys can be obtained by following the instructions at \url{https://github.com/menmo/eliteprospects-api-documentation}
}
\seealso{
\code{\link{get_api_key}} for getting the API key

Other API functions: \code{\link{get_api_key}}
}
