% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_persistent.R
\name{set_user_agent}
\alias{set_user_agent}
\title{Set the User Agent String}
\usage{

  set_user_agent(ua = "EliteProspects R API http://github.com/pbulsink/EliteProspectsR")
}
\arguments{
\item{ua}{The User Agent string to use for API calls. Default is "EliteProspects R API http://github.com/pbulsink/EliteProspectsR"}
}
\description{
Set a User Agent string to notify EliteProspects what service is using the API.
Defaults to "EliteProspects R API http://github.com/pbulsink/EliteProspectsR".
}
