% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{simple_get}
\alias{simple_get}
\title{Base simple get handler}
\usage{
simple_get(path = NULL, id = NULL, filters = NULL, sorts = NULL,
  fields = NULL, limit = 0, offset = 0)
}
\arguments{
\item{path}{The path to get from}

\item{id}{The specific ID to get}

\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}
}
\value{
the results of the built API call
}
\description{
Base simple get handler
}
