% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{simple_list}
\alias{simple_list}
\title{basic list request handler}
\usage{
simple_list(path, filters = NULL, sorts = NULL, fields = NULL,
  limit = 0, offset = 0)
}
\arguments{
\item{path}{The path to get a list of}

\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}
}
\value{
the results of the built API call
}
\description{
basic list request handler
}
