% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_backend.R
\name{simple_search}
\alias{simple_search}
\title{Base simple search handler}
\usage{
simple_search(path = NULL, query = NULL, types = NULL, filters = NULL,
  sorts = NULL, fields = NULL, limit = 0, offset = 0, minScore = 0)
}
\arguments{
\item{path}{The path to get from}

\item{query}{The specific query to search for}

\item{types}{Any limit on type to return}

\item{filters}{Any filters to apply}

\item{sorts}{Any sorting to apply to results}

\item{fields}{Specific fields to retrieve}

\item{limit}{Maximum number of results returned}

\item{offset}{Offset of returned results}

\item{minScore}{Minimum score of search results}
}
\value{
the results of the built API call
}
\description{
Base simple search handler
}
