% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EliteProspectsR-package.R
\docType{package}
\name{EliteProspectsR-package}
\alias{EliteProspectsR}
\alias{EliteProspectsR-package}
\title{EliteProspectsR: Package for scrapping data from eliteprospects.com}
\description{
Utility functions for scrapping various data from eliteprospects.com. This includes draft lists, player stats, player information, league information, and team information. Note that there is an IP-related rate limit, so will likely run into 403 errors when running lots of queries back-to-back.
}
\author{
\strong{Maintainer}: Matthew Palmeri \email{matthew.thomas.palmeri@gmail.com}

}
\keyword{internal}
